+++ samples/snippets/downloadDocumentWithHistory.js
@@ -1,0 +1,34 @@
+'use strict';
+
+const signnow = require('@signnow/api-client')({
+  credentials: 'BASE64_ENCODED_CLIENT_CREDENTIALS',
+  production: true, // if false then uses eval server
+});
+const downloadDocument = signnow.document.download;
+
+const id = 'DOCUMENT_ID_GOES_HERE';
+const token = 'YOUR_ACCESS_TOKEN';
+const options = { withHistory: true }; // option points on downloading document with history
+
+/**
+ * @param {Binary} res - binary data (file) of the entire document
+ */
+const handleResponse = res => {
+  console.log(res);
+};
+
+const handleError = err => {
+  console.error(err);
+};
+
+downloadDocument({
+  id,
+  token,
+  options,
+}, (err, res) => {
+  if (err) {
+    handleError(err);
+  } else {
+    handleResponse(res);
+  }
+});