+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added optional ability to download document with history in *Download Document* feature
+- Updated documentation of *Download Document* feature
+- Updated example of download document in [README](https://github.com/signnow/SignNowNodeSDK/blob/master/README.md#download-document)
+
 ## [v1.5.0] - Unreleased Phase 4
 
 ### Added